Implement bases of the class PdfeContentsStream.
The goal PdfeContentsStream is to be the main interface to handle,
modify and save back PDF contents stream. It should be sufficiently
light so that it can be kept in memory, even on large documents.
The current implementation only focus on bases, i.e. constructors,
PdfeContentsStream::Node definition, doubly-linked nodes list that
represents the stream and deep copy of a contents stream.