Add contents stream to PRGPage.
The PdfeContentsStream is supposed to be heart of contents stream
analysis. Therefore it is directly integrated to the class PRGPage.
Currently, only basic operations are allowed: loading, clearing. The
representation of the stream can be obtained through PRGPage::contents().

At present, no cache system for page contents is implemented. It might be
necessary in the future since loading a large PDF document can use up to
150 mbytes.