Some improvements in PdfeGraphicOperator.h
Add global functions PdfeGOperator::category and PdfeGOperator::str that
return respectively the category and the string representation given an
operator type (PdfeGOperator::Enum).

Improve the class PdfeGraphicOperator. Only keep in memory the type of the
operator, since using previous functions, the category and string can be
efficiently retrieved. Also allows to improve the memory print (can be useful
for PdfeContentsStream).

TODO: Remove depreciated parts and adapt the rest of the code in consequence.