Update PdfeContentsStream behaviour on XObject nodes.
Now store the pointer of the XObject PoDoFo::PdfObject, whatever it is
an image, a form or a PS. It also keeps in memory the type of the XObject.
In the case of a loaded form, a second 'Do' node is added, corresponding to
the closing node of the loaded form's stream.