Draft of implementation of PRPage class.
Its goal is mainly to provide an interface for page handling in a
PDF document. The class is sufficiently flexible so that a PRPage
object can be attached or not to a given PRDocument and PdfPage.

If not, contents/resources and basic properties of the page are stored
locally and modifications can not be saved into a document.
If the page is then attached to a document, it can be synchronised
(using pull/push member functions) with the PDF document itself, so that
modifications made are saved back.

Some content cache features can be implemented using loadContents/clearContents
member functions.

Still need to check for bugs.