Large modifications on the class PdfeResources.
Resources entries are stored independently in PdfeResources members.
It allows easier merging between resources and renaming.
Member functions PdfeResources::load/save are used as interfaces with
real PDF resources objects.

Still need to modify and check the rest of the source code to adapt and
see if there are any regression.