Update PoDoFoExtended classes to the new PdfeResources interface.
Update PoDoFoExtended to the new interface of the PdfeResources class.
In particular, replace the use of push_back by append.
Main modifications appear in PdfeContentsStream::load. When forms from
a stream are loaded, resources objects are appended and a suffix is added
to names in order to avoid any name-conflict between form and parent stream.