Update the class PRPage.
Update the interface of PRPage in order to be ready for asynchronous
use of page contents stream. PRPage is now a sub-class of QObject.
Signals contentsLoaded and contentsCleared have been added. They might be
used by a page contents cache system (e.g. implemented in PRDocument).
When page contents or data is modified, modifications are automatically pushed
to the attached PoDoFo page.