Add a basic implementation of PdfeGElement.
The class PdfeGElement aims to be a common parent of any kind of
graphics element that exists on PDF page and which can be retrieved
from a contents stream (e.g. path, text, image, ...).

Thereby, it contains the basic information which shared by every type
of these elements:
- node ID where is defined in the element in the contents stream;
- graphics state of the stream when it is defined. This information
is optional and can be omitted when not necessary.