Add PdfePathDevice and PdfeBoostStream.
Following the boost::iostreams documentation, the device's member
functions are integrated in a separate class PdfeDataDevice.

PdfeBoostStream is added as a subclass of boost::iostreams, where
text formatting is initialized to comply PDF reference (i.e. fixed
floating precision and english locale).
It is specialized into PdfeDataStream when the device corresponds to
a PdfeDataDevice object.