Update the implementation of PdfeTextWord.
Highly inspired by the previously written PRGTextWord, the most important
features of he class PdfeTextWord have been implemented, i.e.
- basic creation and initialization from a string word or a translation
component;
- simple getters (advance, bbox, unicode string, ...).

Note that in contrary to PRGTextWord, a word object is assumed to be linked
to a text element (unless it is of unknown type). Furthermore, it has sub
node-ID, since it corresponds to a sub-graphical element from a contents stream.